Rydder etter migrering av vilkårsvurdering
