+++ apps/etterlatte-vilkaarsvurdering/src/main/kotlin/Application.kt
@@ -4,6 +4,5 @@ import no.nav.etterlatte.vilkaarsvurdering.config.ApplicationBuilder
 
 fun main() {
     val application = ApplicationBuilder()
-    // application.migrerVilkaarsvurdering()
     application.start()
 }
