+++ apps/etterlatte-vilkaarsvurdering/src/main/resources/db/migration/V8__endre_navn_ny_vilkaarsvurdering_slette_gammel.sql
@@ -1,0 +1,4 @@
+DROP TABLE vilkaarsvurdering;
+
+ALTER TABLE vilkaarsvurdering_v2
+    RENAME TO vilkaarsvurdering;
